requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-cors-headers==3.0.2
django-filter==2.2.0
django-haystack==v3.0b2
djangorestframework==3.10.1
drf-haystack==1.8.6
elasticsearch==5.5.3
elasticstack==0.5.0
idna==2.8
Markdown==3.1.1
mysqlclient==1.4.2
psycopg2
pytz==2019.1
requests==2.22.0
soupsieve==1.9.2
sqlparse==0.3.0
urllib3==1.25.3